+++ Kernel/BootInfo.h
@@ -10,7 +10,7 @@
 #include <Kernel/PhysicalAddress.h>
 #include <Kernel/VirtualAddress.h>
 
-namespace Kernel {
+namespace Kernel::Memory {
 class PageTableEntry;
 }
 
@@ -22,12 +22,12 @@ extern "C" FlatPtr kernel_load_base;
 #if ARCH(X86_64)
 extern "C" u32 gdt64ptr;
 extern "C" u16 code64_sel;
-extern "C" PhysicalAddress boot_pml4t;
 #endif
+extern "C" PhysicalAddress boot_pml4t;
 extern "C" PhysicalAddress boot_pdpt;
 extern "C" PhysicalAddress boot_pd0;
 extern "C" PhysicalAddress boot_pd_kernel;
-extern "C" Kernel::PageTableEntry* boot_pd_kernel_pt1023;
+extern "C" Kernel::Memory::PageTableEntry* boot_pd_kernel_pt1023;
 extern "C" char const* kernel_cmdline;
 extern "C" u32 multiboot_flags;
 extern "C" multiboot_memory_map_t* multiboot_memory_map;
