--- Kernel/Memory/PageDirectory.h
@@ -1,77 +1,0 @@
-/*
- * Copyright (c) 2018-2020, Andreas Kling <kling@serenityos.org>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#pragma once
-
-#include <AK/AtomicRefCounted.h>
-#include <AK/Badge.h>
-#include <AK/HashMap.h>
-#include <AK/IntrusiveRedBlackTree.h>
-#include <AK/RefPtr.h>
-#include <Kernel/Forward.h>
-#include <Kernel/Locking/Spinlock.h>
-#include <Kernel/Memory/PhysicalPage.h>
-
-namespace Kernel::Memory {
-
-class PageDirectory final : public AtomicRefCounted<PageDirectory> {
-    friend class MemoryManager;
-
-public:
-    static ErrorOr<NonnullLockRefPtr<PageDirectory>> try_create_for_userspace();
-    static NonnullLockRefPtr<PageDirectory> must_create_kernel_page_directory();
-    static LockRefPtr<PageDirectory> find_current();
-
-    ~PageDirectory();
-
-    void allocate_kernel_directory();
-
-    FlatPtr cr3() const
-    {
-#if ARCH(X86_64)
-        return m_pml4t->paddr().get();
-#else
-        return m_directory_table->paddr().get();
-#endif
-    }
-
-    bool is_cr3_initialized() const
-    {
-#if ARCH(X86_64)
-        return m_pml4t;
-#else
-        return m_directory_table;
-#endif
-    }
-
-    AddressSpace* address_space() { return m_space; }
-    AddressSpace const* address_space() const { return m_space; }
-
-    void set_space(Badge<AddressSpace>, AddressSpace& space) { m_space = &space; }
-
-    RecursiveSpinlock<LockRank::None>& get_lock() { return m_lock; }
-
-    // This has to be public to let the global singleton access the member pointer
-    IntrusiveRedBlackTreeNode<FlatPtr, PageDirectory, RawPtr<PageDirectory>> m_tree_node;
-
-private:
-    PageDirectory();
-    static void register_page_directory(PageDirectory* directory);
-    static void deregister_page_directory(PageDirectory* directory);
-
-    AddressSpace* m_space { nullptr };
-#if ARCH(X86_64)
-    RefPtr<PhysicalPage> m_pml4t;
-#endif
-    RefPtr<PhysicalPage> m_directory_table;
-    RefPtr<PhysicalPage> m_directory_pages[512];
-    RecursiveSpinlock<LockRank::None> m_lock {};
-};
-
-void activate_kernel_page_directory(PageDirectory const& pgd);
-void activate_page_directory(PageDirectory const& pgd, Thread* current_thread);
-
-}
